fix: use a creator function
