+++ backend/api/core/init.py
@@ -1,4 +1,4 @@
-from fastapi import FastAPI
+from fastapi import FastAPI, Request
 
 from api.core.db import get_connection_pool
 from api.core.settings import settings
@@ -11,14 +11,15 @@ class Database:
 
 def create_app() -> FastAPI:
     app = FastAPI()
-    db = Database()
+    app.state.db = Database()
 
     @app.on_event("startup")
     async def startup():
-        await db.create_pool()
+        await app.state.db.create_pool()
 
     @app.on_event("shutdown")
     async def shutdown():
+
         # cleanup
         pass
 
@@ -26,4 +27,30 @@ def create_app() -> FastAPI:
 
     app.include_router(api_router, prefix=settings.V1_PREFIX)
 
+    @app.get("/")
+    async def read_root():
+        return {"Hello": "World"}
+
+    @app.get("/healthz")
+    async def read_root():
+        return {"whazza": "yolo"}
+
     return app
+
+
+async def connection(request: Request):
+    async with request.app.state.db.pool.acquire() as db:
+        yield db
+
+
+async def transaction(request: Request):
+    async with request.app.state.db.pool.acquire() as db:
+        txn = db.transaction()
+        await txn.start()
+        try:
+            yield db
+        except:  # noqa
+            await txn.rollback()
+            raise
+        else:
+            await txn.commit()
